requirements.txt
requests
undetected-chromedriver
selenium
python-dotenv
pyinstaller
